Fixing up DeviceInfo formatting and tests.
This updates the serialization of DeviceInfo to match the new
formatting. Primarily, this means the splitting of DeviceInfo from one
map with all entries into an array of two maps. The VerifiedDeviceInfo
portion contains entries added in the HAL implementation which are
included as AAD into the signature over the MAC key. The
UnverifiedDeviceInfo is anything added by the app itself whose
authenticity cannot be cryptographically verified.

Bug: 189018262
Test: atest RemoteProvisionerUnitTests
Change-Id: Ifaf1f1dc4d1fc9f65d75bcbcff77c769c286f58e