Fixing notifyKeyGenerated for hybrid RKP systems.
notifyKeyGenerated() is called whenever an attestation key is generated
by an application. Currently, it only checks to see if there are any
unassigned keys in the keypool for the relevant KM implementation.
This change modifies that behavior to also check the total number of
keys in the keypool.

If the total is zero, then provisioning should be
skipped. Zero RKP keys on the system imply that it is a hybrid
factory/remotely provisioned device that has fallen back to factory
provisioning.

Bug: 190645703
Test: atest CtsKeystoreTestCases on a hybrid system
Change-Id: Icaad413bb2e077c99cf3eed83d8bbbfe290149d6
Merged-In: Icaad413bb2e077c99cf3eed83d8bbbfe290149d6